Move generated glean code into a static library
