+++ tests/qml/CMakeLists.txt
@@ -43,13 +43,6 @@ target_sources(qml_tests PRIVATE
     ${MZ_SOURCE_DIR}/apps/vpn/filterproxymodel.h
     ${MZ_SOURCE_DIR}/apps/vpn/glean/glean.h
     ${MZ_SOURCE_DIR}/apps/vpn/glean/glean.cpp
-    ${MZ_SOURCE_DIR}/apps/vpn/glean/generated/metrics.h
-    ${MZ_SOURCE_DIR}/apps/vpn/glean/generated/pings.h
-    ${MZ_SOURCE_DIR}/apps/vpn/glean/private/event.h
-    ${MZ_SOURCE_DIR}/apps/vpn/glean/private/event.cpp
-    ${MZ_SOURCE_DIR}/apps/vpn/glean/private/ping.h
-    ${MZ_SOURCE_DIR}/apps/vpn/glean/private/ping.cpp
-    ${MZ_SOURCE_DIR}/apps/vpn/glean/metrictypes.h
     ${MZ_SOURCE_DIR}/apps/vpn/inspector/inspectorhandler.h
     ${MZ_SOURCE_DIR}/apps/vpn/models/featuremodel.cpp
     ${MZ_SOURCE_DIR}/apps/vpn/models/featuremodel.h
