+++ vpnglean/glean_parser_ext/templates/cpp_metrics_source.jinja2
@@ -9,7 +9,7 @@ Jinja2 template is not. Please file bugs! #}
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 #include "glean/metrictypes.h"
-#include "glean/generated/metrics.h
+#include "glean/generated/metrics.h"
 
 #include <QString>
 #include <QObject>
@@ -24,26 +24,3 @@ Jinja2 template is not. Please file bugs! #}
   {% endfor %}
 {% endfor %}
 
-{% for category_name, objs in all_objs.items() %}
-class __DONOTUSE__{{ category_name|Camelize }} final {
-  Q_GADGET
-
-{% for obj in objs.values() %}
-  Q_PROPERTY({{ obj|type_name }} {{obj.name|camelize }} READ {{ obj.name|camelize }} CONSTANT);
-{% endfor %}
-
-  public:
-    __DONOTUSE__{{ category_name|Camelize }}() = default;
-    ~__DONOTUSE__{{ category_name|Camelize }}() = default;
-
-    {% for obj in objs.values() %}
-    {{ obj|type_name }} {{obj.name|camelize }} () const { return m_{{obj.name|camelize }}; };
-    {% endfor %}
-
-  private:
-    {% for obj in objs.values() %}
-    {% if obj|attr("allowed_extra_keys_with_types")|length > 0 %}__DONOTUSE__{{ obj.name|Camelize }}ExtraParser m_{{obj.name|camelize }}ExtraParser;{%- endif %}
-    {{ obj|type_name }} m_{{obj.name|camelize }} = {{ obj|type_name }}({{obj|metric_id}}{% if obj|attr("allowed_extra_keys_with_types")|length > 0 %}, &m_{{obj.name|camelize }}ExtraParser {%- endif %});
-    {% endfor %}
-};
-{% endfor %}
