+++ vpnglean/glean_parser_ext/templates/macros.jinja2
@@ -91,5 +91,6 @@ FfiExtra __DONOTUSE__{{ obj.name|Camelize }}ExtraParser::fromStruct(const EventM
   ffiExtras.keys.resize(count);
   return ffiExtras;
 }
+
 {% endif %}
 {%- endmacro -%}
