+++ vpnglean/src/event.cpp
@@ -2,9 +2,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-#include "glean/private/event.h"
+#include "glean/event.h"
 
-#include "logger.h"
 #if not(defined(MZ_WASM) || defined(BUILD_QMAKE))
 #  include "vpnglean.h"
 #endif
@@ -15,10 +14,6 @@
 #if defined(UNIT_TEST)
 #  include <QJsonDocument>
 #endif
-
-namespace {
-Logger logger("VPNGlean::EventMetric");
-}  // namespace
 
 EventMetric::EventMetric(int id, EventMetricExtraParser* parser)
     : m_id(id), m_parser(parser) {}
@@ -41,8 +36,8 @@ void EventMetric::record(const QJsonObject& extras) {
                        ffiExtras.keys.size());
 #endif
   } else {
-    logger.error() << "Attempted to record an event with extras, but no extras "
-                      "were provided. Ignoring.";
+    qWarning() << "Attempted to record an event with extras, but no extras "
+                  "were provided. Ignoring.";
     // TODO: record an error.
   }
 }
@@ -59,8 +54,8 @@ void EventMetric::record(const EventMetricExtra& extras) {
                        ffiExtras.keys.size());
 #endif
   } else {
-    logger.error() << "Attempted to record an event with extras, but no extras "
-                      "were provided. Ignoring.";
+    qWarning() << "Attempted to record an event with extras, but no extras "
+                  "were provided. Ignoring.";
     // TODO: record an error.
   }
 }
